requirements.txt
flask==2.3.3
opencv-python==4.8.1.78
numpy<2.0.0
torch
torchvision==0.15.2
facenet-pytorch==2.5.3
pillow
psycopg2-binary==2.9.7
gdown